Adding a 'Create a new User' feature
Driven by a feature spec, new users can now click a button and sign up
with their name and email in a form. Once the form is submitted, there
is a success notice upon redirection to the homepage. The next step
would be to add validation of uniqueness of name and email.